requirements.txt
anyascii==0.3.3
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
babel==2.17.0
beautifulsoup4==4.13.5
binaryornot==0.4.4
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.12.14
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
crispy-bootstrap4==2025.6
cryptography==46.0.1
cssselect2==0.8.0
DAWG-Python==0.7.2
DAWG2==0.13.2
DAWG2-Python==0.9.0
defusedxml==0.7.1
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.0.14
django-appconf==1.1.0
django-autocomplete-light==3.12.1
django-ckeditor==6.7.3
django-ckeditor-5==0.2.15
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-debug-toolbar==6.0.0
django-extensions==4.1
django-filter==25.1
django-import-export==4.3.10
django-js-asset==3.1.2
django-model-utils==5.0.0
django-modelcluster==6.4
django-nested-admin==4.1.4
django-permissionedforms==0.1
django-select2==8.4.1
django-stubs-ext==5.2.2
django-taggit==6.1.0
django-tasks==0.8.1
django-treebeard==4.7.1
djangorestframework==3.16.1
docopt==0.6.2
docxcompose==1.4.0
docxtpl==0.20.1
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
fonttools==4.60.0
freetype-py==2.5.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
Jinja2==3.1.6
laces==0.1.2
lxml==6.0.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
pillow==11.3.0
pillow_heif==1.1.0
pycairo==1.28.0
pycparser==2.23
pydyf==0.11.0
Pygments==2.19.1
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy3==2.0.4
pymorphy3-dicts-ru==2.4.417150.4580142
pypdf==6.1.1
pyperclip==1.11.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.1.1
python-monkey-business==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
reportlab==4.4.4
requests==2.32.3
rich==14.0.0
rlPyCairo==0.4.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
striprtf==0.0.29
svglib==1.6.0
tablib==3.8.0
telepath==0.3.1
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritools==5.0.0
urllib3==2.3.0
wagtail==7.1.1
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
Willow==1.11.0
xhtml2pdf==0.2.17
xlrd==2.0.2
zopfli==0.2.3.post1
